simplify Loading Patch from URL
- able to simplify after removing non-IdentifierCorrections.
