+++ Sources/iTunes/Patchable+URL.swift
@@ -7,18 +7,6 @@
 
 import Foundation
 
-extension Array where Element: Codable {
-  static fileprivate func load(from url: URL) throws -> Self {
-    try load(from: try Data(contentsOf: url, options: .mappedIfSafe))
-  }
-
-  static fileprivate func load(from data: Data) throws -> Self {
-    let decoder = JSONDecoder()
-    decoder.dateDecodingStrategy = .iso8601
-    return try decoder.decode(Self.self, from: data)
-  }
-}
-
 extension Patchable {
   func createPatch(_ fileURL: URL) throws -> Patch {
     switch self {
@@ -26,11 +14,7 @@ extension Patchable {
       .replaceComments, .replaceDateReleased, .replaceAlbumTitle, .replaceYear,
       .replaceTrackNumber, .replaceIdSongTitle, .replaceIdDiscCount, .replaceIdDiscNumber,
       .replaceArtist, .replacePlay:
-      try Self.identifierCorrections(fileURL)
+      try Patch.load(fileURL)
     }
-  }
-
-  static func identifierCorrections(_ fileURL: URL) throws -> Patch {
-    Patch.identifierCorrections(try Array<IdentifierCorrection>.load(from: fileURL))
   }
 }
